AArch64: Fix implicit conversion from ptrdiff_t to int
In common/aarch64/loopfilter-prim.cpp there is a function call to abs()
with the result of subtracting two pointer types, which results in an
implicit cast from ptrdiff_t to int as abs() is only defined for int
prior to C++11 (and x265 uses gnu++98).

Make conversion from ptrdiff_t to int explicit by using a cast in
common/aarch64/loopfilter-prim.cpp.

+++ source/common/aarch64/loopfilter-prim.cpp
@@ -158,7 +158,7 @@ static void processSaoCUE2_neon(pixel *rec, int8_t *bufft, int8_t *buff1, int8_t
 {
     int x;
 
-    if (abs(buff1 - bufft) < 16)
+    if (abs(static_cast<int>(buff1 - bufft)) < 16)
     {
         for (x = 0; x < width; x++)
         {
